Added Image Processing + the environment key json
ImageProcess currently just peels out all the text, but that's about it. no information yet, need to actually process the text afterwards honestly

The json key is there for the text to actually work; download it, save it in a nice safe folder, and specify in your terminal 
"export GOOGLE_APPLICATION_CREDENTIALS="[YOUR_FILE_PATH_HERE]" for it to work. Do that every time you open the terminal (if you close it).